Use lock on homeassistant level
