+++ hassio/utils/__init__.py
@@ -1,13 +1,30 @@
 """Tools file for HassIO."""
 from datetime import datetime
+import logging
 import re
 
+_LOGGER = logging.getLogger(__name__)
 RE_STRING = re.compile(r"\x1b(\[.*?[@-~]|\].*?(\x07|\x1b\\))")
 
 
 def convert_to_ascii(raw):
     """Convert binary to ascii and remove colors."""
     return RE_STRING.sub("", raw.decode())
+
+
+def process_lock(method):
+    """Wrap function with only run once."""
+    async def wrap_api(api, *args, **kwargs):
+        """Return api wrapper."""
+        if api.lock.locked():
+            _LOGGER.error(
+                "Can't excute %s while a task is in progress", method.__name__)
+            return False
+
+        async with api.lock:
+            return await method(api, *args, **kwargs)
+
+    return wrap_api
 
 
 class AsyncThrottle(object):
